requirements.txt
streamlit
pandas
numpy
folium
streamlit-folium
Pillow
